requirements.txt
streamlit>=1.30
pdfplumber>=0.10
python-docx>=1.1
openpyxl>=3.1
pydantic>=2.5
icalendar>=5.0
reportlab>=4.0
requests>=2.31
beautifulsoup4>=4.12
pandas>=2.0
pytest>=7.4
pyinstaller>=6.3
